feat(type-safe-api): add mock integration for returning hardcoded responses
This adds `Integrations.mock` for hardcoding a response for API operations. The response will
automatically include CORS headers if cors was enabled on the API. Note that this isn't type safe to
the operation's response types, however this provides a foundation for generating type safe mocks as
part of the generated cdk infrastructure package"
