+++ packages/type-safe-api/scripts/generators/python/templates/operationConfig.mustache
@@ -49,6 +49,7 @@ OperationLookup = {
     "{{operationId}}": {
         "path": "{{path}}",
         "method": "{{httpMethod}}",
+        "contentTypes": [{{^hasConsumes}}"application/json"{{/hasConsumes}}{{#consumes}}{{#mediaType}}"{{{.}}}",{{/mediaType}}{{/consumes}}]
     },
 {{/operation}}
 {{/operations}}
