+++ packages/type-safe-api/scripts/generators/typescript/templates/operationConfig.mustache
@@ -48,6 +48,7 @@ export const OperationLookup = {
     {{nickname}}: {
         path: '{{path}}',
         method: '{{httpMethod}}',
+        contentTypes: [{{^consumes}}'application/json'{{/consumes}}{{#consumes}}{{#mediaType}}'{{{.}}}',{{/mediaType}}{{/consumes}}],
     },
     {{/operation}}
     {{/operations}}
