+++ packages/type-safe-api/src/construct/integrations/index.ts
@@ -1,3 +1,6 @@
 /*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0 */
+export * from "./integration";
 export * from "./integrations";
+export * from "./lambda";
+export * from "./mock";
