+++ packages/type-safe-api/test/resources/specs/single.yaml
@@ -98,6 +98,26 @@ paths:
             application/x-mpegurl:
               schema:
                 type: string
+  /multiple-content-types:
+    post:
+      operationId: multipleContentTypes
+      requestBody:
+        required: true
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/TestRequest'
+          application/pdf:
+            schema:
+              type: string
+              format: binary
+      responses:
+        '200':
+          description: Success
+          content:
+            application/json:
+              schema:
+                type: string
 components:
   schemas:
     ApiError:
